requirements.txt
# Buttplug.io Python client (requires Intiface Central to be running)
buttplug-py

# eyeware-beam is NOT on PyPI.
# Install it manually from the Beam SDK — see README.md for instructions.
# eyeware-beam

mss
Pillow
 
# NSFW detection (AdamCodd/vit-base-nsfw-detector)
# torch / torchvision: install the CPU build if you have no NVIDIA GPU:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# For CUDA (GPU) builds visit https://pytorch.org/get-started/locally/
torch
torchvision
transformers
